endpoint: Split endpoint cache into cache.go
During compilation, it's useful to grab a cache of the relevant endpoint
info that can be used by the next layer of code to generate the BPF
datapath, then release the lock on the endpoint to allow other work to
continue. This cache facilitates this, but may end up holding
stale information about an endpoint which may be subsequently deleted.

This patch just refactors the code related to this cache into its own
file for better code organisation.
