+++ pkg/endpoint/bpf.go
@@ -252,80 +252,6 @@ func hashHeaderfile(hashWriter hash.Hash, filepath string) (hash.Hash, error) {
 	return hashWriter, nil
 }
 
-// epInfoCache describes the set of lxcmap entries necessary to describe an Endpoint
-// in the BPF maps. It is generated while holding the Endpoint lock, then used
-// after releasing that lock to push the entries into the datapath.
-// Functions below implement the EndpointFrontend interface with this cached information.
-type epInfoCache struct {
-	// revision is used by the endpoint regeneration code to determine
-	// whether this cache is out-of-date wrt the underlying endpoint.
-	revision uint64
-
-	// For lxcmap.EndpointFrontend
-	keys  []*lxcmap.EndpointKey
-	value *lxcmap.EndpointInfo
-
-	// For datapath.loader.endpoint
-	epdir    string
-	id       string
-	ifName   string
-	endpoint *Endpoint // Used to get the endpoint's logger.
-}
-
-// Must be called when endpoint is still locked.
-func (e *Endpoint) createEpInfoCache(epdir string) *epInfoCache {
-	ep := &epInfoCache{
-		revision: e.nextPolicyRevision,
-		endpoint: e,
-		epdir:    epdir,
-		id:       e.StringID(),
-		ifName:   e.IfName,
-		keys:     e.GetBPFKeys(),
-	}
-
-	var err error
-	ep.value, err = e.GetBPFValue()
-	if err != nil {
-		log.WithField(logfields.EndpointID, e.ID).WithError(err).Error("getBPFValue failed")
-		return nil
-	}
-	return ep
-}
-
-// InterfaceName returns the name of the link-layer interface used for
-// communicating with the endpoint.
-func (ep *epInfoCache) InterfaceName() string {
-	return ep.ifName
-}
-
-// StringID returns the endpoint's ID in a string.
-func (ep *epInfoCache) StringID() string {
-	return ep.id
-}
-
-// Logger returns the logger for the endpoint that is being cached.
-func (ep *epInfoCache) Logger() *logrus.Entry {
-	return ep.endpoint.Logger()
-}
-
-// StateDir returns the directory for the endpoint's (next) state.
-func (ep *epInfoCache) StateDir() string {
-	return ep.epdir
-}
-
-// GetBPFKeys returns all keys which should represent this endpoint in the BPF
-// endpoints map
-func (ep *epInfoCache) GetBPFKeys() []*lxcmap.EndpointKey {
-	return ep.keys
-}
-
-// GetBPFValue returns the value which should represent this endpoint in the
-// BPF endpoints map
-// Must only be called if init() succeeded.
-func (ep *epInfoCache) GetBPFValue() (*lxcmap.EndpointInfo, error) {
-	return ep.value, nil
-}
-
 // addNewRedirectsFromMap must be called while holding the endpoint lock for
 // writing. On success, returns nil; otherwise, returns an error  indicating the
 // problem that occurred while adding an l7 redirect for the specified policy.
